==> notebooks/1. Data preparation.ipynb <==
import numpy as np
import pandas as pd

from tqdm.notebook import tqdm
from sklearn.model_selection import train_test_split

# Первичный анализ данных

data_path = '../data/full_data.csv'

df = pd.read_csv(data_path, index_col=0)

df.head()

Оценим соотношение классов.

df.groupby('is_duplicate').count()

Выборка сильно несбалансирована.

Посмотрим, присутсвует ли что-то кроме латиницы.

u0 = map(lambda x: list(x), df['name_1'].tolist())
total = []
for lst in tqdm(u0, total=df['name_1'].shape[0]):
    total.extend(lst)

u1 = map(lambda x: list(x), df['name_2'].tolist())
for lst in tqdm(u1, total=df['name_2'].shape[0]):
    total.extend(lst)

set(total)

Стоит убрать мусорные символы типа ", #, % и так далее. В эти символы также можно включить цифры и скобки:

trash_symbols = {
    '"', '#', '%', '&', "'", '(', ')', '*', '+', ',', '-', '.', '/',
    '0', '1', '2', '3', '4', '5', '6', '7', '8', '9',
    ':', ';', '<', '>', '?', '@', '[', '\\', ']', '`', '{', '\xa0', '«',
    '±', '»', '\u0e00', '‘', '’', '\u3000', '。', '﹠', '＆', '（', '）',
    '，'
}

Также в фразах присутсвует несколько языков:
* Латиница
* Кирилица
* Иероглифы
* Прочие

Разделим их и посмотрим на соотношение языков в датасете.

latin = {
    'A', 'B', 'C', 'D', 'E', 'F', 'G', 'H', 'I', 'J', 'K', 'L', 'M',
    'N', 'O', 'P', 'Q', 'R', 'S', 'T', 'U', 'V', 'W', 'X', 'Y', 'Z',
    'a', 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i', 'j', 'k', 'l', 'm',
    'n', 'o', 'p', 'q', 'r', 's', 't', 'u', 'v', 'w', 'x', 'y', 'z'
}

cyrillic = {
    'Ё', 'А', 'Б', 'В', 'Г', 'Д', 'Е', 'Ж', 'З', 'И', 'Й', 'К', 'Л', 'М',
    'Н', 'О', 'П', 'Р', 'С', 'Т', 'У', 'Ф', 'Х', 'Ц', 'Ч', 'Ш', 'Щ', 'Ы',
    'Ь', 'Э', 'Ю', 'Я', 'а', 'б', 'в', 'г', 'д', 'е', 'з', 'и', 'к', 'л',
    'м', 'н', 'о', 'п', 'р', 'с', 'т', 'у', 'ш', 'ь', 'э'
}

unknown_1 = {
    '¿', 'Ã', 'Ç', 'Í', 'Ñ', 'Ó', 'Ö', 'Ú', 'Ü', 'á', 'ç', 'è', 'é', 'í',
    'ñ', 'ó', 'õ', 'ö', 'ú', 'ü', 'ę', 'İ', 'Ł', 'ł', 'ő', 'Ş', 'ű'
}

unknown_2 = {
    'ر', 'س', 'ف', 'ك', 'م', 'و', 'ي'
}

unknown_3 = {
    '上', '东', '京', '份', '会', '公', '北', '双', '司', '团', '式', '彤',
    '技', '料', '新', '方', '日', '有', '术', '材', '株', '水', '海', '社',
    '程', '股', '虹', '防', '限', '集', '雨'
}

(set(total)
 - latin
 - cyrillic
 - unknown_1
 - unknown_2
 - unknown_3
 - trash_symbols)

Будем смотреть, какие языки охватывает каждая пара наименований.

def classify_language(
    string: str,
    languages_base: list,
    trash_symbols: set
) -> list:
    answer = []
    letters = set(string) - trash_symbols - {' '}
    for language in languages_base:
        answer.append(bool(language & letters) + 0)
    return answer

languages_base = [
    latin,
    cyrillic,
    unknown_1,
    unknown_2,
    unknown_3
]

languages_base_names = [
    'latin',
    'cyrillic',
    'unknown_1',
    'unknown_2',
    'unknown_3'
]

print(df['name_1'].iloc[0])
classify_language(df['name_1'].iloc[0], languages_base, trash_symbols)

df_copy = df.copy()

u = df_copy[['name_1', 'name_2']].apply(
    lambda row: classify_language(
        set(row['name_1']) | set(row['name_2']),
        languages_base,
        trash_symbols
    ) ,
    axis=1)

u.head(3)

u2t = pd.DataFrame(u.tolist(), index=u.index, columns=languages_base_names)

u2t.head(3)

df_copy[languages_base_names] = u2t[languages_base_names]

df_copy.head(3)

Частота языков:

# В целом
df_copy.groupby('is_duplicate')[languages_base_names].sum()

# Пары с одним языком
one_language_df = df_copy[df_copy[languages_base_names].sum(axis=1) == 1]
one_language_df.groupby('is_duplicate')[languages_base_names].sum()

# Исходное число пар
df.groupby('is_duplicate').count()

В целом можно заключить, что взяв пары только с латинскими символами, мы не сильно потеряем от общей выборки. Учитывая то, насколько потенциально обременительно может быть тащить за собой все языки, это скорее правильное решение.

# Пары с различными языками
multi_language_df = df_copy[df_copy[languages_base_names].sum(axis=1) > 1]
print(multi_language_df.shape)
multi_language_df.groupby('is_duplicate')[languages_base_names].sum()

df[['name_1', 'name_2']].duplicated().any()

# Очистка выборки

# Возьмем фразы только на латинском языке
target_columns = ['name_1', 'name_2', 'is_duplicate']
latin_df = one_language_df[one_language_df['latin'] == 1][target_columns]

def clear_string_from_symbols(string: str, symbols: set) -> str:
    res = ''.join(
        filter(
            lambda symb: symb not in symbols,
            list(string)
        )
    )
    return res

print(latin_df.iloc[25]['name_1'])
r = clear_string_from_symbols(latin_df.iloc[25]['name_1'], trash_symbols)
print(r)

latin_df_copy = latin_df.copy()

# Уберем мусорные символы
for col in ['name_1', 'name_2']:
    latin_df_copy[col] = latin_df_copy[col].apply(
        lambda val: clear_string_from_symbols(val, trash_symbols)
    )

latin_df.head()

latin_df_copy.head()

latin_df_copy.shape

latin_df_copy.groupby('is_duplicate').count()

latin_df_copy.isna().any()

latin_df_copy.head(3)

latin_df_copy.to_csv('clear_latin_data.csv', index=False)

# Разбиение данных на обучение и тест

Разделим выюорку на train и test сохранив процентное соотношение классов. Для теста выделим $10\%$ всей выборки.

ds_path = '../data/clear_latin_data.csv'
df_clear = pd.read_csv(ds_path)

df_clear.head(3)

X, y = df_clear[['name_1', 'name_2']].values, df_clear[['is_duplicate']].values

X_train, X_test, y_train, y_test = train_test_split(
    X, y, stratify=y, shuffle=True, test_size=0.1, random_state=681)

print(f'X_train: {X_train.shape}')
print(f'X_test: {X_test.shape}')
print()
print(f'y_train: {y_train.shape}')
print(f'y_test: {y_test.shape}')

y_train_coeff = round(
    y_train.sum() / (y_train.shape[0] - y_train.sum()), 4)
print(f'y_train: {y_train_coeff}')

y_test_coeff = round(
    y_test.sum() / (y_test.shape[0] - y_test.sum()), 4)
print(f'y_test: {y_test_coeff}')

Сохраним данные

df_train_data = pd.DataFrame(X_train, columns=['name_1', 'name_2'])
df_train_data['is_duplicate'] = y_train

df_test_data = pd.DataFrame(X_test, columns=['name_1', 'name_2'])
df_test_data['is_duplicate'] = y_test

df_train_data.to_csv('train_data.csv', index=False)

df_test_data.to_csv('test_data.csv', index=False)

# Проверка данных

train_data_path = '../data/train_data.csv'
test_data_path = '../data/test_data.csv'

train_data = pd.read_csv(train_data_path)
test_data = pd.read_csv(test_data_path)

print(train_data.shape)
train_s = train_data['is_duplicate'].sum()
round(train_s / (train_data['is_duplicate'].shape[0] - train_s), 4)

print(test_data.shape)
test_s = test_data['is_duplicate'].sum()
round(test_s / (test_data['is_duplicate'].shape[0] - test_s), 4)
