==> notebooks/2. Metrics.ipynb <==
import numpy as np
import pandas as pd
from sklearn.metrics import precision_score, recall_score

train_data_path = '../data/train_data.csv'
test_data_path = '../data/test_data.csv'

train_data = pd.read_csv(train_data_path)
test_data = pd.read_csv(test_data_path)

test_data.head(3)

test_data.shape

На нужен Recall, так как он отслеживает отрицательные ответы модели:

$$Precision = \frac{TP}{TP + FP}$$

$$Recall = \frac{TP}{TP + FN}$$

y_true = test_data['is_duplicate']
y_pred = np.append(np.full(test_data.shape[0]-1, 0), [1])

def eval_metrics(y_true, y_pred):
    precision = precision_score(y_true, y_pred)
    recall = recall_score(y_true, y_pred)
    return precision, recall

precision, recall = eval_metrics(y_true, y_pred)

print(f'precision: {precision}')
print(f'recall: {recall}')
