==> notebooks/3. Custom bagging.ipynb <==
import random

import numpy as np
import pandas as pd

train_data_path = '../data/train_data.csv'
test_data_path = '../data/test_data.csv'

train_data = pd.read_csv(train_data_path)
test_data = pd.read_csv(test_data_path)

# Вариант бэггинга

train_data.shape

train_data.head(3)

count_classes_df = train_data.groupby('is_duplicate').count()
count_classes_df

positive_n = count_classes_df.loc[1, 'name_1']
negative_n = count_classes_df.loc[0, 'name_1']

negative_n // positive_n

bag_size = int(positive_n*1.5)
bag_size

negative_pairs_df = train_data[train_data['is_duplicate'] == 0]

bags = 10
random.seed = 456

bags_indexes = np.array([
    random.sample(range(negative_n), bag_size) for _ in range(bags)
])

bags_indexes.shape
